requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1
torchvision==0.14.1
diffusers[torch]==0.19.3
xformers==0.0.16
transformers>=4.25.1
bitsandbytes==0.35.4
decord==0.6.0
pytorch-lightning<2
omegaconf==2.2.3
nerfacc==0.3.3
trimesh==3.9.8
pyhocon==0.3.57
icecream==2.1.0
PyMCubes==0.1.2
accelerate
modelcards
einops
ftfy
piq
matplotlib
opencv-python
imageio
imageio-ffmpeg
scipy
pyransac3d
torch_efficient_distloss
tensorboard
rembg
segment_anything
gradio==3.50.2
libigl
open_clip_torch
fire
